WIP: gpm_to_wrl (#129)
* WIP: Added a notebook in which we can put together the matching.
* ADD: io_func.py with `read_gpm-function, `read_trmm`-function and `read_gr`-function 
* ADD: add gpm code to wradlib codebase
* FIX:: several fixes for numpy 1.12 
* WIP: update docstrings, refactoring (#128)
* DOC: revise DOCS
* MAINT: clean-up CI scripts, reordering, switching as appropriate to reduce CI-Time
* MAINT: add python 3.6 to travis-matrix

+++ wradlib/io.py
@@ -1061,7 +1061,7 @@ def read_gamic_scan_attributes(scan, scan_type):
         ele_start = np.round(ray_header['elevation_start'], 1)
         ele_stop = np.round(ray_header['elevation_stop'], 1)
         angle_step = np.round(sattrs['angle_step'], 1)
-        angle_step = np.round(sattrs['ele_stop'], 1) / angle_step
+        angle_step = int(np.round(sattrs['ele_stop'], 1) / angle_step)
         # Elevation corresponding to 1st ray
         if ele_start[0] < 0:
             ele_start = ele_start[1:]
